pyproject.toml
[tool.poetry]
name = "listinha"
version = "0.1.0"
description = ""
authors = ["Raul_Coelho"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
whitenoise = "^6.4.0"
pillow = "^9.5.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.2"
django-cors-headers = "^3.14.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
